+++ source/isaaclab/isaaclab/sensors/ray_caster/ray_caster.py
@@ -228,6 +228,8 @@ class RayCaster(SensorBase):
         self._data.pos_w = torch.zeros(self._view.count, 3, device=self._device)
         self._data.quat_w = torch.zeros(self._view.count, 4, device=self._device)
         self._data.ray_hits_w = torch.zeros(self._view.count, self.num_rays, 3, device=self._device)
+        if self.cfg.track_ray_distance:
+            self._data.ray_distance = torch.full((self._view.count, self.num_rays), self.cfg.max_distance, device=self._device, dtype=torch.float32)
 
     def _update_buffers_impl(self, env_ids: Sequence[int]):
         """Fills the buffers of the sensor data."""
@@ -293,12 +295,16 @@ class RayCaster(SensorBase):
 
         # ray cast and store the hits
         # TODO: Make this work for multiple meshes?
-        self._data.ray_hits_w[env_ids] = raycast_mesh(
+        self._data.ray_hits_w[env_ids], ray_distance, _, _ = raycast_mesh(
             ray_starts_w,
             ray_directions_w,
             max_dist=self.cfg.max_distance,
             mesh=self.meshes[self.cfg.mesh_prim_paths[0]],
-        )[0]
+            return_distance=self.cfg.track_ray_distance,
+        )
+
+        if self.cfg.track_ray_distance:
+            self._data.ray_distance[env_ids] = ray_distance.clip(max=self.cfg.max_distance)
 
         # apply vertical drift to ray starting position in ray caster frame
         self._data.ray_hits_w[env_ids, :, 2] += self.ray_cast_drift[env_ids, 2].unsqueeze(-1)